backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
pandas
scipy
scikit-learn
anthropic
python-dotenv
statsmodels
